+++ src/birdnet/models/__init__.py
@@ -1,1 +1,2 @@
+from birdnet.models.model_v2m4_protobuf import ModelV2M4Protobuf as ModelV2M4
 from birdnet.models.model_v2m4_tflite import ModelV2M4TFLite