+++ src/birdnet/models/model_v2m4_protobuf.py
@@ -1,6 +1,8 @@
 import os
 import zipfile
+from logging import getLogger
 from pathlib import Path
+from typing import List, Optional, cast
 
 import numpy as np
 import numpy.typing as npt
@@ -56,7 +58,7 @@ class DownloaderProtobuf():
 
   def _download_model_files(self) -> None:
     dl_path = "https://tuc.cloud/index.php/s/ko6DA29EMwLBe3c/download/BirdNET_v2.4_protobuf.zip"
-    dl_size = None
+    dl_size = 124524452
     self._version_path.mkdir(parents=True, exist_ok=True)
 
     zip_download_path = self._version_path / "download.zip"
@@ -74,6 +76,18 @@ class DownloaderProtobuf():
       assert self._check_model_files_exist()
 
 
+def try_get_gpu_otherwise_return_cpu() -> tf.config.LogicalDevice:
+  all_gpus = tf.config.list_logical_devices('GPU')
+  if len(all_gpus) > 0:
+    first_gpu = all_gpus[0]
+    return first_gpu
+  all_cpus = tf.config.list_logical_devices('CPU')
+  if len(all_cpus) == 0:
+    raise Exception("No CPU found!")
+  first_cpu = all_cpus[0]
+  return first_cpu
+
+
 class ModelV2M4Protobuf(ModelV2M4Base):
   """
   Model version 2.4
@@ -81,7 +95,7 @@ class ModelV2M4Protobuf(ModelV2M4Base):
   This class represents version 2.4 of the model.
   """
 
-  def __init__(self, /, *, language: Language = "en_us") -> None:
+  def __init__(self, /, *, language: Language = "en_us", custom_device: Optional[str] = None) -> None:
     """
     Initializes the ModelV2M4 instance.
 
@@ -90,6 +104,8 @@ class ModelV2M4Protobuf(ModelV2M4Base):
     language : Language, optional, default="en_us"
         The language to use for the model's text processing. Must be one of the following available languages:
         "en_us", "en_uk", "sv", "da", "hu", "th", "pt", "fr", "cs", "af", "uk", "it", "ja", "sl", "pl", "ko", "es", "de", "tr", "ru", "no", "sk", "ar", "fi", "ro", "nl", "zh".
+    custom_device : str, optional, default=None
+        This parameter allows specifying a custom device on which computations should be performed. If custom_device is not specified (i.e., it has the default value None), the program will attempt to use a GPU (e.g., /device:GPU:0) by default. If no GPU is available, it will fall back to using the CPU. By specifying a device string such as /device:GPU:0 or /device:CPU:0, the user can explicitly choose the device on which operations should be executed.
 
     Raises:
     -------
@@ -98,8 +114,27 @@ class ModelV2M4Protobuf(ModelV2M4Base):
     """
     super().__init__(language=language)
 
+    device: tf.config.LogicalDevice
+    if custom_device is None:
+      device = try_get_gpu_otherwise_return_cpu()
+    else:
+      matched_device = None
+      available_devices: List[tf.config.LogicalDevice] = tf.config.list_logical_devices()
+
+      for logical_device in available_devices:
+        if logical_device.name == custom_device:
+          matched_device = logical_device
+          break
+      if matched_device is None:
+        raise ValueError(
+          f"Device '{custom_device}' doesn't exist. Please select one of the following existing device names: {', '.join(d.name for d in available_devices)}.")
+      device = matched_device
+    self._device = device
+
+    logger = getLogger(__name__)
+    logger.info(f"Using device: {self._device.name}")
+
     model_folder = self._model_version_folder / "Protobuf"
-    # model_folder = Path("/home/mi/code/birdnet/src/birdnet_debug/BirdNET_v2.4_gpu")
     downloader = DownloaderProtobuf(model_folder)
     downloader.ensure_model_is_available()
 
@@ -114,14 +149,15 @@ class ModelV2M4Protobuf(ModelV2M4Base):
 
   def _predict_species(self, batch: npt.NDArray[np.float32]) -> npt.NDArray[np.float32]:
     assert batch.dtype == np.float32
-    with tf.device('/device:CPU:1'):
+    with tf.device(self._device):
       prediction: Tensor = self._audio_model.basic(batch)["scores"]
     prediction_np: npt.NDArray[np.float32] = prediction.numpy()
     return prediction_np
 
   def _predict_species_location(self, sample: npt.NDArray[np.float32]) -> npt.NDArray[np.float32]:
     assert sample.dtype == np.float32
-    prediction: Tensor = self._meta_model(sample)
+    with tf.device(self._device):
+      prediction: Tensor = self._meta_model(sample)
     prediction = tf.squeeze(prediction)
     prediction_np: npt.NDArray[np.float32] = prediction.numpy()
     return prediction_np