+++ src/birdnet_tests/conftest.py
@@ -1,5 +1,7 @@
 import logging
 
+import tensorflow as tf
+
 
 def pytest_configure():
   loggers = {"numba"}
@@ -7,3 +9,10 @@ def pytest_configure():
     logger = logging.getLogger(l)
     logger.disabled = True
     logger.propagate = False
+
+  gpus = tf.config.list_physical_devices('GPU')
+  for gpu in gpus:
+    try:
+      tf.config.experimental.set_memory_growth(gpu, True)
+    except RuntimeError as e:
+      print(e)